scenarios/loan-mlflow-sklearn/pyproject.toml
[project]
name = "data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.24",
    "joblib>=1.5.3",
    "mlflow>=3.8.1",
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "scikit-learn>=1.8.0",
    "wandb>=0.16.0",
    "clearml>=1.13.0",
    "matplotlib>=3.10.0",
    "PyYAML>=6.0.1",
    "venturalitica-sdk",
    "fairlearn>=0.10.0",
    "torch>=2.0.0",
    "torchvision>=0.24.1",
]
[tool.uv.sources]
venturalitica-sdk = { path = "../../packages/venturalitica-sdk", editable = true }
